move DeadNPC interface to lib
