+++ src/Castle/createDungeon.d.ts
@@ -1,7 +1,4 @@
-import { Building } from '../../lib/buildings/_common'
-import { DeadNPC } from 'src/NPCGeneration/setupDeath'
-import { Namesake, NPC } from '../../lib/npc-generation/_common'
-import { Town } from '../../lib/town/_common'
+import { Building, DeadNPC, Namesake, NPC, Town } from '@lib'
 
 interface Setup {
   createDungeon(town: Town, opts: any): Dungeon
